Continued refactor, refs #577
